Move REST API spec to a separate project
REST API is defined using JSON files base on Swagger spec.
They are currently located at a dir under mgmt project.

This force anyone who wish to use the API specs to pull the entire mgmt project.
Users will need only the JSON files to build, for example a Ruby client.

This is a first patch of two which will move mgmt/api to a new osv-rest-api repository.
In this patch files are added to the new repository
In the next one, mgmt/api directory will be replaced with a submodule.

Signed-off-by: Tzach Livyatan <tzach@cloudius-systems.com>
Signed-off-by: Pekka Enberg <penberg@iki.fi>